Fix ConcurrentModificationException in GVRAnimationEngine onDrawFrame
If an animation is started or stopped in the course of either a
GVROnFinish.finish() or a GVROnRepeat.iteration(), a
ConcurrentModificationException will result, either when
iterator.remove() is called or on the next iteration.  The most
likely scenario will be starting a new animation as part of a
finish() implementation: this is an easy way to chain multiple simple
animations together to create a longer, more involved animation.

This change copies the animation queue and then executes on that
copy, allowing animations to be added to or removed from the original
queue without risk of a CME.