Fix the 'leaky texture' problem.
This looked at first like a reference counting problem.
That turned out to be an artifact of the debugging code,
and the real problem was with the 'object identity' code:

1) The 'Dereference thread' was only waking up every 1000
   registered allocations. The Note 4's GL hardware only
   only has room for about 490 1024x1024 textures.

2) Once we removed the sWrappers reference to wrappers
   with a use count of 1, it could be quite a while until
   Java garbage collection actually freed up the wrappers.

The solution is to wake the 'Dereference thread' more often,
and to explicitly recycle() any GVRRecyclableObject (on the
GL thread!) immediately after dereferencing.

The GVRConsole is a bit inefficiently implemented, creating
a new texture every frame instead of updating the existing
texture, but it now runs for tens of thousands of frames
without problem.