GVRBitmapTexture::Constructor - Invalid parameter message is missing (null Bitmap)
For this issue, there are two ways to show the message.
First one is on Java side, we can just modify the GVRBitmapTexture class to throw an InvalidParameterException when we find that
the input Bitmap is null. But it requires us to add at least one method since this class uses super() and crash happen inside
super().

Another way to do that is to print an error message in logcat with gvrf tag. Firstly, there is already an series of error handle
in native side, it is simpler to add one more condition at native side. Secondly I think we need to encourage developers to make
 more use of logcat information with tag gvrf.

This patch takes the second way.