Fix a memory leak in the 'object identity' code.
The basic GVRF C++->Java design is that every getX() call creates
a new std::shared_ptr that wraps a pointer to a native object
and increments a reference count. The getX() call then casts
the pointer to the std::shared_ptr to a long, and returns it
to Java, which wraps a new Java instance around it. Since
essentially all the Java instance's methods work by making
JNI calls to the native object, it wasn't supposed to matter
that there might be several different wrapper for the same
native object.

Except, there are cases where it does. People who actually
understand object-oriented programming do wild, wacky things
like 'subclassing'. Strange stuff, like declaring classes
that extend GVRSceneObject and that add fields with important
app data in them. Then they get very upset when, say, eye
picking into the scene graph gives them a scene object that
acts just like the one they put in ... but isn't actually the
same instance, or even the same derived class.

Sarcasm aside, this lack of 'object identity' was a major bug
at the beginning. I fixed this by adding a mapping
from native pointers to wrapper classes, so that users could
get back the actual instances they put in. But I did make a
few mistakes of my own in the process: with any luck, this is
the last of them.

The GVRHybridObject.wrapper(long) method is called just before
creating a new wrapper for the new std::shared_ptr. It returns
the existing wrapper, if any. The various Java getX() methods
call a factory(GVRContext, long) method, instead of directly
calling the (GVRContext, long) constructor; the factory()
method calls GVRHybridObject.wrapper(), and either returns
the (appropriately cast) non-null result, or returns a brand
new wrapper class.

The bug here is that the original design expected every new
std::shared_ptr to end up inside a new wrapper. Each
std::shared_ptr would end up getting freed when the Java
wrapper got finalized; the use counts *would* correspond
to the number of native + Java references to an object. But!
My GVRHybridObject.wrapper() code simply ignored its ptr
argument if it found that it referred to an existing object:
the std:shared_ptr was leaked, and the Java getUseCount()
just kept climbing.

You can easily see this in, say, the Solar System sample,
which calls getMainCameraRig() on the main scene object
several times. It's always the same GVRScene instance; it's
always the same GVRCameraRig; but the use count is higher at
the end of onInit() than at the start.

This fix simply calls NativeHybridObject.delete() on the
std::shared_ptr if it refers to an already wrapped object.
This frees the memory and decrements the reference count,
just as if we *did* create a new wrapper which was eventually
finalized. With this change, things seem to work as before -
except that repeated calls to, say, getMainCameraRig() do
not result in an inflated use count.

This has been a VERY long commit message for a very small
change, because the 'shared_ptr abuse' is really complicated
*and* really central. I'm fairly certain that this fix is
both right and safe; I do want my bright, meticulous coworkers
to do their best to spot any flaws!