Add Java stack trace when native error is thrown. And add java stack trace when GVRCameraRig.getVecK() uses an invalid key
In current GVRf, the life cycle is controlled by VR thread which is a C-thread, so we can not simply do JNIEnv->throwNew
to trigger a Java exception and get current Java stack trace. This patch provides a class GVRJniException simple global call
for JNI codes to print current Java stack trace especially when native error is going to be thrown. The tag for this stack trace
is gvrf. It is because currently all errors reported in our project uses tag "gvrf". I think we need to encourage developers to
use this tag to debug GearVRf app.

There is an issue that the GVRF will simply cause a native crash without any information when GVRCameraRig.getVecX() uses a key
that isn't there. The patch simply uses the GVRJniException to print stack trace, and then throw the error.