Fix a 'mandelbug' in async loading.
The root cause of the bug was using the same GVRAndroidResource in
multiple load calls. This is now illegal, as it is problematic in
another way, as well: If the later request(s) can't be consolidated
with the earlier request(s) (because the earlier request(s) have
already been completed) the later request(s) will reload the
resource ... using a closed stream.

The proximate cause of the bug was the way that the
private FutureResource class (in GVRAsynchronousResourceLoader)
was using the user-supplied GVRAndroidResource for
synchronization. This has now been corrected, so that each
FutureResource uses a unique, private Object for synchronization.

Also cuts a bit of log spam, and fixes one or two minor issues
I ran into while tracking down the mandelbug.

GearVRf-DCO-1.0-Signed-off-by: Jon Shemitz
j.shemitz@samsung.com