libjav8.so updates
The current version of libjav8.so was built with NDK 16b. It is reported to crash
when used in sxr-x3ddemo on armeabi. I spent much time investigating this issue
however I cannot explain why this crash shall occur.

These updated binaries are produced in the same build with NDK 15c. The armeabi
version has been verified to work with sxr-x3ddemo.