Fix lighting regressions (#69)
* Asset Loader updates main scene on GL thread

When replacing the main scene, the asset loader now modifies the scene on the GL thread. This fixes the problem with lights not being added when replacing the entire scene. They were being added and then cleared on the GL thread because lights are cleared only on that thread. Now we add the new nodes on the GL thread as well guaranteeing that new lights are added after old lights are cleared.

* include lights with parents that have no names

Fixes regression caused by animation rework

* fix typo