Fix intermittent crash on bitmap update (#77)
* Fix for the TestSendEvents test - controller class sending events now to the right target (app)
Renamed IActivityEvents to the more appropriate IApplicationEvents (since the app is the target for these)

* Add left&right screenshot support to the monoscopic backend

* Fix intermittent crash on bitmap update

It is possible that an image may request update before the rendering thread has attached Java. This causes a NULL JNIEnv pointer which will crash. This patch programs around that by using the same JNIEnv pointer throughout the life of the lock and checking if it is null. If the rendering thread has not yet attached to JNI, the update will be skipped until next frame.