Adding IAssetImportEvents, initialize mesh deformer default weights (#89)
* Fix parsing of GLB files with embedded  buffers

This PR just updates the assimp library binaries, the fixes were in assimp

* Adding IAssetImportEvents

IAssetEvents is limiting because it does not provide the SXRTexture which failed to load in IAssetEvents.onTextureError. If this is provided, handlers can easily substitute textures for those that fail to load.

* Jassimp: set default blend weights

Set the default blend weights for the mesh morph component

* Adding IAssetImportEvents, initialize mesh deformer

IAssetEvents is limiting because it does not provide the SXRTexture which failed to load in IAssetEvents.onTextureError. IAssetImportEvents eliminates this problem. Ideally IAssetEvents should have worked that way...
Jassimp now sets the default blend weights for the SXRMeshMorph component from the assimp input meshes.

* Fix bug with opening assets from /storage/emulated/...

* Fix problem with asset event handler

* Remove avatar load handler after loading

* Fix event target for avatar animation events

* fix x3d build problem

* support multiview in X3D shader