Fixes and updates to support AR Pet demo (#95)
* Added rigid body reset feature

This feature is intended to adapt the rigid body to owner's transform
when it is modified "outside" physics symulation. Optionally physics
collider can be rebuilt if needed (in case scene object's collider is
changed).

* API to config continuous collision threshold

* Fix physics simulation for custom scene

The physics simulation was not starting after scene's exchange.
If the user want to pause the simulation it needs disable the SXRWorld
component and enable again to resume the simulation.