Support multiple bind poses on a single skeleton
The bind pose that used to be kept in the skeleton is now kept in the SXRSkin component associated with the mesh. This allows the same skeleton to be used with multiple meshes with different bind poses.
Removed SXRSkeleton.getBindPose, SXRSkeleton.setBindPose
Added SXRSkin.setInverseBindPose to set the bind pose for the skin
Modified asset loader to use new APIs
Made Java SXRSkeleton ande C++ Skeleton more thread-safe by locking around updates
SXRPoseMapper will not animate if either skeleton is disabled