Remove glGetAttribLocation, roll gl_program into gl_shader
1. Modified GLShader to include the shader creation logic in GLProgram
2. Leave layout(location=xxx) in the GL shaders
3. Bind vertex attributes base on vertex descriptor of shader
   (this is done between compile and link phases)
3. Remove glGetAttribLocation calls in GLVertexBuffer

Our shaders already explicitly declare the locations for vertex attributes because it is needed by Vulkan. Previously these declarations were removed before being submitted to GL for compilation. Now they are left in and the code assumes the shader author matches these in the Java shader. (If they don't, the shader won't work on Vulkan anyway. Now it might not work on GL either).

We can't do the same thing for uniforms because GLSL doesn't allow location specifiers here (yet).