Added name to SXRAnimation
When a morph  animation is imported with its geometry, the morphs are attached to the appropriate SXRMeshMorph component. If the file contains only the morph animations and no geometry, there is no way to indicate which node the morph should change. By adding a name, the morph animation can be named after the node it affects.

This allows the morph animation to be apploied to a target node from a different file.