new assimp libs with support for sparse accessors (#137)
* fixed issues with scaling (meters to centimeters)

The original code scaled the entire matrix, not just the positions. If you are already working in centimeters, your objects do not need more scaling. This PR simplifies the ARCore version of mixed reality and fixes the problems with scaling.

* remove debug printouts

* new assimp libs with support for sparse accessors

Some GLTF files which have morphing in them use sparse accessors. We support them in assimp now.

* remove debug prints

* Add api to IAvatarService interface

* reset destination pose if skeleton changes during animation