Fix AR passthru scaling when in headset
The current implementation fills the screen with the video passthru when using a headset. This is actually wrong - the video should be the same size as it would be viewing through the headset. In this case, the video does not fill the entire screen as the camera FOV is 60 and VR is 90.

The ARCore implementation of passthru on a headset has been changed to correctly size the video. Now it is at real scale but no longer fills the screen.

Also made a bug fix to clearAvatar to support changing avatar geometry